pyproject.toml
[tool.poetry]
name = "bobbins"
version = "0.0.1alpha1"
description = "Bot for managing help posts in a forum."
authors = ["Zech Zimmerman <hi@zech.codes>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.12"
hikari = "^2.0.0.dev115"
hikari-lightbulb = "^2.3.1"
sqlalchemy = "2.0.0rc2"
psycopg = "^3.1.7"
greenlet = "^2.0.1"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
